Version 1.0.3
Added RCmdPresetServer command that takes in a server name to be used for all the operations. Until you call the same command passing an empty string "", you will not prompted to select the server, which makes it easy to execute repeated remote commands in the same remote session.

Also added two new commands \sf and \sb to search forward and backward for the current word in the remote session.